# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Triplet API",
  "description": "Trip planning aplication API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/trips/{trip_id}/links": {
   "post": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Create Link",
    "operationId": "create_link_trips__trip_id__links_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SavedLinkCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedLinkResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Get Links",
    "operationId": "get_links_trips__trip_id__links_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/SavedLinkResponse"
         },
         "title": "Response Get Links Trips  Trip Id  Links Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/links/screenshot": {
   "post": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Create Screenshot",
    "description": "Save a screenshot, like a post or a map someone sent, and read the places in it.",
    "operationId": "create_screenshot_trips__trip_id__links_screenshot_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_screenshot_trips__trip_id__links_screenshot_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedLinkResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/links/{link_id}": {
   "get": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Get Link",
    "operationId": "get_link_trips__trip_id__links__link_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Link Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedLinkResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Update Link",
    "operationId": "update_link_trips__trip_id__links__link_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Link Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SavedLinkUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedLinkResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Delete Link",
    "operationId": "delete_link_trips__trip_id__links__link_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Link Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/links/{link_id}/refresh": {
   "post": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Refresh Link",
    "operationId": "refresh_link_trips__trip_id__links__link_id__refresh_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Link Id"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SavedLinkResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/trips/{trip_id}/links/{link_id}/activity": {
   "post": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Add Link To Itinerary",
    "operationId": "add_link_to_itinerary_trips__trip_id__links__link_id__activity_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "trip_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Trip Id"
      }
     },
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Link Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LinkToActivity"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActivityResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/link-images/{link_id}": {
   "get": {
    "tags": [
     "Saved Links"
    ],
    "summary": "Get Link Image",
    "operationId": "get_link_image_link_images__link_id__get",
    "parameters": [
     {
      "name": "link_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Link Id"
      }
     },
     {
      "name": "sig",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Sig"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ActivityResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "trip_id": {
      "type": "integer",
      "title": "Trip Id"
     },
     "source_link_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Link Id"
     },
     "place_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "location": {
      "type": "string",
      "title": "Location"
     },
     "latitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Latitude"
     },
     "longitude": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Longitude"
     },
     "start_time": {
      "type": "string",
      "format": "date-time",
      "title": "Start Time"
     },
     "end_time": {
      "type": "string",
      "format": "date-time",
      "title": "End Time"
     },
     "estimated_cost": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Estimated Cost"
     }
    },
    "type": "object",
    "required": [
     "id",
     "trip_id",
     "title",
     "location",
     "start_time",
     "end_time"
    ],
    "title": "ActivityResponse"
   },
...
   "Body_create_screenshot_trips__trip_id__links_screenshot_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_create_screenshot_trips__trip_id__links_screenshot_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LinkToActivity": {
    "properties": {
     "place_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Id"
     },
     "title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 5000
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "location": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "start_time": {
      "type": "string",
      "format": "date-time",
      "title": "Start Time"
     },
     "end_time": {
      "type": "string",
      "format": "date-time",
      "title": "End Time"
     },
     "estimated_cost": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 99999999.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Estimated Cost"
     }
    },
    "type": "object",
    "required": [
     "start_time",
     "end_time"
    ],
    "title": "LinkToActivity"
   },
...
   "SavedLinkCreate": {
    "properties": {
     "url": {
      "type": "string",
      "maxLength": 2083,
      "minLength": 1,
      "format": "uri",
      "title": "Url"
     },
     "place_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 5000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "url"
    ],
    "title": "SavedLinkCreate"
   },
   "SavedLinkResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "trip_id": {
      "type": "integer",
      "title": "Trip Id"
     },
     "added_by_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Added By Id"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "platform": {
      "type": "string",
      "title": "Platform"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "custom_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Custom Title"
     },
     "author_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Author Name"
     },
     "thumbnail_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Thumbnail Url"
     },
     "place_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "caption": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Caption"
     },
     "summary": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summary"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "processed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Processed At"
     },
     "places": {
      "items": {
       "$ref": "#/components/schemas/ExtractedPlaceResponse"
      },
      "type": "array",
      "title": "Places",
      "default": []
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "trip_id",
     "url",
     "platform",
     "status",
     "created_at"
    ],
    "title": "SavedLinkResponse"
   },
   "SavedLinkUpdate": {
    "properties": {
     "place_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Place Name"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 5000
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "custom_title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Custom Title"
     }
    },
    "type": "object",
    "title": "SavedLinkUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}